+++ src/legacy/core_plugins/data/public/query_bar/components/query_bar_input.test.mocks.ts
